+++ examples/usual/example26.rb
@@ -13,7 +13,7 @@ module Usual
     private
 
     def assign_first_id
-      self.first_id = inputs.ids[0]
+      outputs.first_id = inputs.ids[0]
     end
   end
 end