+++ examples/usual/example3.rb
@@ -6,18 +6,25 @@ module Usual
     input :middle_name, type: String, required: false
     input :last_name, type: String
 
+    internal :prepared_full_name, type: String
+
     output :full_name, type: String
 
+    make :prepare_full_name
     make :assign_full_name
 
     private
 
-    def assign_full_name
-      self.full_name = [
-        inputs.first_name,
-        inputs.middle_name,
-        inputs.last_name
+    def prepare_full_name
+      int.prepared_full_name = [
+        inp.first_name,
+        inp.middle_name,
+        inp.last_name
       ].compact.join(" ")
+    end
+
+    def assign_full_name
+      out.full_name = int.prepared_full_name
     end
   end
 end