+++ examples/usual/example35.rb
@@ -11,15 +11,15 @@ module Usual
     private
 
     def assign_number_5
-      self.number = 5
+      outputs.number = 5
     end
 
     def assign_number_6
-      self.number = 6
+      outputs.number = 6
     end
 
     def assign_number_7
-      self.number = 7
+      outputs.number = 7
     end
   end
 end