+++ examples/usual/example49.rb
@@ -19,11 +19,11 @@ module Usual
     private
 
     def assign_number_6
-      self.number = 6
+      outputs.number = 6
     end
 
     def assign_number_7
-      self.number = 7
+      outputs.number = 7
     end
   end
 end