+++ examples/wrong/example5.rb
@@ -11,7 +11,7 @@ module Wrong
     private
 
     def assign_invoice_number
-      self.invoice_number = inputs.invoice_number
+      outputs.invoice_number = inputs.invoice_number
     end
   end
 end