+++ lib/servactory/internals/collection.rb
@@ -5,7 +5,7 @@ module Servactory
     class Collection
       # NOTE: http://words.steveklabnik.com/beware-subclassing-ruby-core-classes
       extend Forwardable
-      def_delegators :@collection, :<<, :each, :map, :merge
+      def_delegators :@collection, :<<, :each, :map, :merge, :find
 
       def initialize(*)
         @collection = Set.new
@@ -13,6 +13,10 @@ module Servactory
 
       def names
         map(&:name)
+      end
+
+      def find_by(name:)
+        find { |internal| internal.name == name }
       end
     end
   end