+++ website/i18n/ru/docusaurus-plugin-content-docs/current/CHANGELOG.md
@@ -8,4 +8,16 @@ pagination_label: Changelog
 
 # Changelog
 
-Soon
+## Не выпущено
+
+### Изменено
+
+- Изменена работа с `input`, `output` и `internal` атрибутами ([#76](https://github.com/afuno/servactory/pull/76))
+
+### Улучшено
+
+- Улучшена производительность ([#76](https://github.com/afuno/servactory/pull/76))
+
+## 1.6.14
+
+First release in the changelog